+++ crates/daemon/src/control_plane_server/serve.rs
@@ -1,4 +1,5 @@
 use super::*;
+use std::sync::atomic::AtomicU64;
 
 fn build_control_plane_routes() -> Router<ControlPlaneHttpState> {
     Router::new()
